# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type StripInputPlacement = typeof StripInputPlacement[keyof typeof StripInputPlacement];


export const StripInputPlacement = {
  front: 'front',
  back: 'back',
} as const;

export type StripInputFilter = typeof StripInputFilter[keyof typeof StripInputFilter];


export const StripInputFilter = {
  mono: 'mono',
  sepia: 'sepia',
} as const;

export interface StripInput {
  /**
     * @minItems 1
     * @maxItems 4
     * @items.maxLength 5000000
     */
  photos: string[];
  /** @maxLength 180 */
  note: string;
  placement: StripInputPlacement;
  filter: StripInputFilter;
}

export type StripSummaryPlacement = typeof StripSummaryPlacement[keyof typeof StripSummaryPlacement];


export const StripSummaryPlacement = {
  front: 'front',
  back: 'back',
} as const;

export type StripSummaryFilter = typeof StripSummaryFilter[keyof typeof StripSummaryFilter];


export const StripSummaryFilter = {
  mono: 'mono',
  sepia: 'sepia',
} as const;

export interface StripSummary {
  id: string;
  createdAt: string;
  photoCount: number;
  note: string;
  placement: StripSummaryPlacement;
  filter: StripSummaryFilter;
}

export type Strip = StripSummary & {
  photos: string[];
  shareUrl?: string;
};

export interface Error {
  error: string;
}