# lib/api-zod/src/generated/types/strip.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StripSummary } from './stripSummary';

export type Strip = StripSummary & {
  photos: string[];
  shareUrl?: string;
};