# lib/api-zod/src/generated/types/stripInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StripInputFilter } from './stripInputFilter';
import type { StripInputPlacement } from './stripInputPlacement';

export interface StripInput {
  /**
     * @minItems 1
     * @maxItems 4
     * @items.maxLength 5000000
     */
  photos: string[];
  /** @maxLength 180 */
  note: string;
  placement: StripInputPlacement;
  filter: StripInputFilter;
}