# lib/api-zod/src/generated/types/stripSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StripSummaryFilter } from './stripSummaryFilter';
import type { StripSummaryPlacement } from './stripSummaryPlacement';

export interface StripSummary {
  id: string;
  createdAt: Date;
  photoCount: number;
  note: string;
  placement: StripSummaryPlacement;
  filter: StripSummaryFilter;
}